… experiments/exp_20250904_192729/README.md
… experiments/exp_20250904_192729/code/__pycache__/data_loader.cpython-310.pyc
… experiments/exp_20250904_192729/code/__pycache__/derp_vae.cpython-310.pyc
… experiments/exp_20250904_192729/code/__pycache__/enhanced_data_loader.cpython-310.pyc
… experiments/exp_20250904_192729/code/__pycache__/enhanced_derp_vae.cpython-310.pyc
… experiments/exp_20250904_192729/code/analysis.py
… experiments/exp_20250904_192729/code/enhanced_data_loader.py
… experiments/exp_20250904_192729/code/enhanced_derp_vae.py
… experiments/exp_20250904_192729/code/enhanced_experiment.py
… experiments/exp_20250904_192729/code/test_implementation.py
… experiments/exp_20250904_192729/code/train.py
… experiments/exp_20250904_192729/results/enhanced_experiment_results.json
… experiments/exp_20250904_192729/results/experiment_analysis.md
… experiments/exp_20250904_192729/results/experiment_results.png
… experiments/exp_20250904_192729/results/small_scale_results.json
… section_notes/05-experiment-runs.md